Cleanup plugins structure
